+++ app/js/app.js
@@ -2,9 +2,14 @@
 
 
 // Declare app level module which depends on filters, and services
-angular.module('myApp', ['myApp.filters', 'myApp.services', 'myApp.directives', 'ngI18n']).
+var myApp = angular.module('myApp', ['myApp.filters', 'myApp.services', 'myApp.directives', 'ngI18n']).
   config(['$routeProvider', function($routeProvider) {
     $routeProvider.when('/view1', {templateUrl: 'partials/partial1.html', controller: MyCtrl1});
     $routeProvider.when('/view2', {templateUrl: 'partials/partial2.html', controller: MyCtrl2});
     $routeProvider.otherwise({redirectTo: '/view1'});
   }]);
+
+myApp.value('ngI18nConfig', {
+    //without leading and trailing slashes
+    basePath: "app/i18n"
+});