+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-v0.1.0
+### v0.2.0
 
-First release (basic functionality
+* Added global config basePath
+
+### v0.1.0
+
+* First release (basic functionality)